+++ codemp/rd-vanilla/tr_surfacesprites.cpp
@@ -726,7 +726,7 @@ static void RB_DrawVerticalSurfaceSprites(shaderStage_t* stage, shaderCommands_t
 				fb = posj + randomchart[randomindex] * step;
 				randomindex += randominterval;
 
-				rightvectorcount = rightvectorcount + 1 & 3;
+				rightvectorcount = static_cast<byte>((static_cast<int>(rightvectorcount) + 1) & 3);
 
 				if (fa > 1.0)
 					continue;
